layout with justand init
