+++ package.json
@@ -18,7 +18,8 @@
     "react-icons": "^4.2.0",
     "react-image-gallery": "^1.2.7",
     "react-vertical-timeline-component": "^3.3.3",
-    "styled-components": "^5.3.1"
+    "styled-components": "^5.3.1",
+    "zustand": "^4.1.1"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.0",
