+++ pages/_app.jsx
@@ -1,7 +1,15 @@
 import "tailwindcss/tailwind.css"
+import Footer from "../src/components/Layout/Footer"
+import Header from "../src/components/Layout/Header"
 
 function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
+  return (
+    <>
+      <Header />
+      <Component {...pageProps} />
+      <Footer />
+    </>
+  )
 }
 
 export default MyApp
