+++ pages/songs.jsx
@@ -1,9 +1,7 @@
 import React, { useState, useEffect } from "react"
-import Header from "../src/components/Header"
 import SongCard from "../src/components/SongCard"
 
 import { ReadDB } from "../src/utility/Firebase"
-import Footer from "../src/components/Footer"
 import LoadingScreen from "../src/components/LoadingScreen"
 
 const recents = () => {
@@ -20,13 +18,11 @@ const recents = () => {
   }
   return (
     <div>
-      <Header current="Songs" />
       <div className="lg:flex flex flex-wrap justify-center align-center ">
         {data.map((item) => {
           return <SongCard item={item} key={item.imgURL} />
         })}
       </div>
-      <Footer />
     </div>
   )
 }
