+++ src/components/Card.jsx
@@ -13,7 +13,6 @@ const Card = ({ text, link, imgPath }) => {
           width={400}
           height={400}
           className="w-full h-21 rounded-lg sm:p-2"
-          // layout="fill"
           src={imgPath}
         />
       </div>
