+++ src/components/MerchCard.jsx
@@ -1,5 +1,4 @@
 import React from "react"
-import Image from "next/image"
 
 const MerchCard = ({ item }) => {
   return (
